Moved role def sp assingment test
